Implement intra-floor indoor routing
This commit implements intra-floor indoor routing. It is based on the
idea that there are fixed navigational "waypoints" on each floor and
that from the set of the origin point, the destination point and the
waypoints, a next point should be chosen that is not already visited
and that is closest to the current point and closer to the destination
than the current point.

This works but produces some undesirable routes (taking the user
through non-navigable areas). The next commit will strictly define which
waypoints are navigable with which other waypoints (waypoints are always
navigable with a feature).